[FIX] Fixed a problem where close() called close(true)
This change is fairly dramatic if a client was depending on the old behavior. However, the concept of unsubscribe has other options so making close call close(false) definitely seems like the right choice, and matches the doc.

Added a test to insure the behavior of close and close(true) is correct.

Fixes #94