[ADDED] Added an option for the error listener and connection listener on the streaming connection, so that it is easier to set those.
I decided not to provide a default listener because the only option is to log
and I didn't want the library to muddy the log space. However, this option makes it
easy to add your own logging.

Also added a test and fixed a timing problem in another test.

Fixes #93